Update main.cs
Fixed bug that had occurred due to misunderstanding of the problem. I had thought the problem was asking to return the longest substring containing characters whose frequency was no more than k, when the problem had been asking to find the longest substring that should only contain k distinct characters. Fortunately, it was a quick fix.